Start project in Actix-web
This project uses Actix-web and builds on a lot of the things learned
from writing the uDevGames website.

As of this commit these things are done:

* Database migrations
* Data models
* Homepage placeholder
* Login/logout workflow
* Site theme (Bootstrap, jQuery{,-ujs})

Yet to be done is the UI work for "snippets," which are the little
blurbs that we like to put on the site to link to cool articles about
game development and to current/past uDevGames jams and contests.